api/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-telegram-bot>=21.0
httpx>=0.26.0
libsql-client>=0.3.0
pyjwt>=2.8.0
bcrypt>=4.1.0
python-multipart>=0.0.6
groq>=0.4.0
matplotlib>=3.8.0
